backend/requirements.txt
langchain
langchain-community
langchain-ollama
faiss-cpu
fastapi
uvicorn
python-dotenv
